Add ability to save normalization params during preprocessing
Previously the normalization parameters for a dataset were calculated each time a generator was instantiated. As these parameters were calculated from a sample size of 100 random images, there could be minor variations in the normalization process between instantiations.

We decided to add the ability to calculate the normalization parameters during the preprocessing stage on a sample size of 1000 images and then save the mean and standard deviations to a json file. Once stored, these values are read for future runs.